Deprecate run_suite.sh.
Robotest configuration is exceedingly complex, and having the shell
script parts of it split between the gravity repo and the robotest repo
makes how variables are used difficult to trace.

With custom robotest app upgrade testing refacoring the gravity side of
robotest configuration, now is an ideal time to revisit the robotest
side as well.